fix dtype conversion

+++ unittest/python/test_user_type.py
@@ -10,7 +10,7 @@ cols = 20
 def test(dtype):
     rng = np.random.default_rng()
     mat = np.array(np.ones((rows, cols)).astype(np.intc), dtype=dtype)
-    mat = rng.random((rows, cols), dtype=dtype)
+    mat = rng.random((rows, cols)).astype(dtype)
     mat_copy = mat.copy()
     assert (mat == mat_copy).all()
     assert not (mat != mat_copy).all()
